Refactor booleans
